pyproject.toml
[project]
name = "svc-mcp-datadog"
version = "0.1.0"
description = "Datadog MCP server with Keycard authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.0",
    "keycardai-mcp-fastmcp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
